--- launch/example_webcam.launch
@@ -1,8 +1,0 @@
- <?xml version="1.0" encoding="utf-8"?>
- 
- <launch>
-   <!-- <include file="$(find realsense2_camera)/launch/rs_camera.launch"/> -->
-   <node pkg="movie_publisher" name="webcam_pub" type="webcam_pub" args=0/>
-   <node pkg="motpy_ros" name="face_tracking" type="face_tracking.py"/>
- 
- </launch>
