+++ motpy_ros/CMakeLists.txt
@@ -1,9 +1,8 @@
 cmake_minimum_required(VERSION 3.0.2)
 project(motpy_ros)
 
-find_package(catkin REQUIRED COMPONENTS
-  rospy
-)
+find_package(catkin REQUIRED COMPONENTS rospy)
+
 
 catkin_package(
 )
@@ -12,3 +11,7 @@ include_directories(
 # include
   ${catkin_INCLUDE_DIRS}
 )
+# add_service_files(
+#   FILES
+#   srv/motpy_bbox.srv
+# )
