+++ motpy_ros/package.xml
@@ -4,20 +4,13 @@
   <version>0.0.0</version>
   <description>The motpy_ros package</description>
 
-  <maintainer email="ubuntu@todo.todo">ubuntu</maintainer>
+  <maintainer email="ray255ar@gmail.com">ar-ray-code</maintainer>
 
-
-  <license>TODO</license>
+  <license>Apache License, Version 2.0</license>
 
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>rospy</build_depend>
   <build_export_depend>rospy</build_export_depend>
   <exec_depend>rospy</exec_depend>
 
-
-  <!-- The export tag contains other, unspecified, tags -->
-  <export>
-    <!-- Other tools can request additional information be placed here -->
-
-  </export>
 </package>
