+++ motpy_ros/srv/motpy_bbox.srv
@@ -1,0 +1,5 @@
+darknet_ros_msgs BoundingBoxes
+
+---
+
+darknet_ros_msgs BoundingBoxes
